fix: rename package to files
