+++ pkg/files/io.go
@@ -1,4 +1,4 @@
-package helpers
+package files
 
 import (
 	"io"
